Segregate action, apply material ui
