+++ src/actions/visibilityFilter.js
@@ -1,0 +1,6 @@
+export const setVisibilityFilter = filter => dispatch =>{
+  return dispatch({
+    type: 'SET_VISIBILITY_FILTER',
+    filter
+  })
+}
