+++ src/containers/firebase/config.js
@@ -1,0 +1,8 @@
+export const firebaseConfig = {
+  apiKey: "AIzaSyC8Vh5FWgLsbPjrinFxWCgkip1TXres2fs",
+  authDomain: "r-r-f-todo.firebaseapp.com",
+  databaseURL: "https://r-r-f-todo.firebaseio.com",
+  projectId: "r-r-f-todo",
+  storageBucket: "r-r-f-todo.appspot.com",
+  messagingSenderId: "544806072099"
+};
