+++ src/reducers/barcode.js
@@ -1,0 +1,12 @@
+const barcode = (state = null, action) => {
+  switch (action.type) {
+    case 'GENERATE_BARCODE':
+      return action.barcode
+    case 'ERASE_BARCODE':
+      return action.barcode
+    default:
+      return state
+  }
+}
+
+export default barcode
