+++ src/reducers/index.js
@@ -4,10 +4,12 @@ import visibilityFilter from './visibilityFilter'
 import auth from './auth'
 import sources from './sources'
 import charges from './charges'
+import barcode from './barcode'
 export default combineReducers({
   todos,
   visibilityFilter,
   auth,
   sources,
-  charges
+  charges,
+  barcode
 })
